# src/docs/openapi.yaml
openapi: 3.0.1
info:
  title: Simple Comercio
  description: Template Service
  contact:
    email: sc@simplecomercio.com
  version: 0.0.1
servers:
  - url: 'http://localhost:4123/api/v2'
    variables: {}
    description: Servidor de desarrollo
...
paths:
  /customer:
    get:
      tags:
        - Customer
      summary: Lista clientes
      operationId: listCustomers
      parameters:
        - in: query
          name: filter
          description: 'Ej: {"id": "60116f6821563333a5895749", "name": "claudio"}'
          schema:
            type: string
      responses:
        '200':
          description: Correcto.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerResponse'
    post:
      tags:
        - Customer
      summary: Crea un cliente
      operationId: createCustomer
      requestBody:
        description: Datos del cliente
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Customer'
        required: true
      responses:
        '201':
          description: '--'
        '400':
          description: '--'
      x-codegen-request-body-name: body
...
components:
  schemas:
    Customer:
      type: object
      properties:
        profileImage:
          type: string
          nullable: true
    CustomerResponse:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        profileImage:
          type: string
          format: binary
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
  securitySchemes:
    ownAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - ownAuth: [] 